https://www.codechef.com/problems/CHEFGARD
All submissions for this problem are available.

There are n trees on a road in Chef's garden. Each tree has one of the three colours out of red, green or blue. You are given information about the colours of the trees by a string s. if s[i] is 'R', it means that ith tree is red. Similarly 'G' denotes green and 'B' denotes blue colored tree.

Chef likes to create a beautiful pattern out of the tree colours. A pattern is called beautiful if all the trees of the same colour are continuous. eg. "RRBGG" and "GGBBB" are beautiful patterns, but "RGRB" is not a beautiful pattern because trees of red color are not continuous.

For achieving beautiful pattern, chef is permitted to change the colours of the trees. For changing the colour of ith tree from any colour to any other colour, he needs to pay a cost of c[i] rupees. Find out the minimum amount of rupees the chef requires, to create a beautiful pattern out of the colours of the n trees.

Input

First line of the input contains a single integer T denoting number of test cases.
Description of T test cases follows.
First line of each test case will contain a single integer n denoting number of trees on the road in chef's garden.
Second line will contain a single line containing string s as defined in the problem.
Third line will contain n space separated integers denoting the array c, i.e. c[1], c[2], .., c[n].
Output

For each test case, print a single line containing a single integer denoting minimum amount of rupees needed to create a beautiful pattern.